Move keypair generation to prepare()
The keypair should only be generated if the parameters have been verified and
all other validation has succeeded. Therefore it's better to move it from
OA2ProxyServlet.verifyAndGet() to MPOA2ProxyServlet.prepare() (the latter class
extends the former). Unfortunately there is no easy place to do this in the
myproxy package, only in the master-portal package.