Improve: don't use unset state and HTTP status
Adding the state parameter for ServletHTTPClientException is not helpful since
it's not known. Note that OpenID requires it to be returned in an AuthN response
error when the client sent one, but then it must match that value. Setting it to
a (none) value hence never makes sense.
Putting the HTTP status in the description is only useful when it has been set,
i.e. when it's >0.