Improve: store the ID token in the OA2Asset
By storing the ID token in the OA2Asset, we don't need to use the static
idTokenStore in ATServer2. The latter one in any case does not work once a token
is renewed, that would require a similar static store in RTServer2. Using the
OA2Asset is much simpler and more flexible.
- We add getter/setter methods to OA2Asset itself.
- appropriate toMap and fromMap entries in OA2AssetConverter, essentially just
  (de)serializing.
- a new key in OA2AssetSerializationKeys.
- make sure to add the ID token in OA2MPService.getAccessToken() and
  OA2MPService.refresh()