Bugfix: prevent NPE when tables are inconsistent
The oa2cli did not properly cleanup the entries in the permissions (and
approvals) tables, hence some of the IDs returned by PermissionsStore's
getClients() and getAdmins() might no longer exist. This resulted in
null-pointer exceptions. We protect against this by producing an error and by
skipping them when creating the list in PermissionServer's listClients() and
listAdmins().
Although this should prevent response.getClients() in ResponseSerializer's
serialize() from returning a list containing some null clients, we also protect
there by producing an error and skipping the invalid entries.