Bugfix: allow admin clients to trigger email
Admin clients weren't able to trigger the NewClientEvent. By generalizing the
Client to a BaseClient (which is a common parent for both (OA2)Client and
AdminClient, we can easily use the same code for both. We only need to properly
adjust for the replacements, in NewClientNotifier and OA2NewClientNotifier. We
introduce a standard String for a field that's not present for an admin client.
The only minor issue is that we cannot use different mail templates for both
type of clients, which would require separate MailUtil instances for both.