Improve: Remove dependency on mysql-connector-java
No need to add an explicit dependency on mysql-connector-java for tomcat
servlets. Note that there are probably a few more maven artefacts that could
have it removed as they also will only run inside tomcat, but this one is
sufficient to prevent problems for us. The cli tools will probably all need it,
as do the tests.