Bugfix: map uses wrong key name iat
The toMap() and fromMap() methods use the OIDC claim name (iat) instead of the
SQL column name (issuedat). This prevents a proper updating of the issuedat
column with the value of the iat claim obtained from the ID token.
We could alternatively have changed the column name to match the claim name, but
that would break backwards compatibility.