Improve: use same column name as in mysql template
The template in oa4mp-client-oauth2/src/main/resources/oauth2-mysql.sql uses
issuedat in all lower case, use the same in the code, such that auto-creating
and creating using the template results in the same column name.
Note that MySQL is (mostly) ignoring the case of the column name.
This improvement is the third of three issues relating to the iat claim/issuedat
column, see also "Bugfix: issuedat is a TIMESTAMP not DATE" and "Bugfix: map
uses wrong key name iat"